feat: add user register function

+++ internal/app/storage/model/user.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type User struct {
 	gorm.Model
-	UserID   int    `gorm:"column:user_id;unique" json:"id"`                // id
+	UserID   int64  `gorm:"column:user_id;unique" json:"id"`                // id
 	NickName string `gorm:"column:nick_name;NOT NULL" json:"nick_name"`     // 昵称
 	Password string `gorm:"column:password;NOT NULL" json:"-"`              // 密码
 	Avatar   string `gorm:"column:avatar;NOT NULL" json:"avatar"`           // 头像
